Rename ConnectionMaxAge filter to ConnectionExpiration
