+++ CONTRIBUTORS.md
@@ -15,3 +15,4 @@
 - Paul Mellor(https://github.com/PaulRMellor)
 - Piotr Płaczek(https://github.com/piotrpdev)
 - Tiberiu Sabau(https://github.com/tibisabau)
+- May Abo-Alrejal(https://github.com/May-Abo)
