+++ kroxylicious-docs/docs/_assemblies/assembly-built-in-filters.adoc
@@ -41,8 +41,8 @@ The Kroxylicious SASL Inspection filter extracts the authenticated principal fro
 and Kafka Broker and makes it available to the other filters in the chain.
 For information on using the filter, see the {SASLInspectionGuide}.
 
-== Connection Max Age filter
+== Connection Expiration filter
 
-The Kroxylicious Connection Max Age filter closes client connections after a configurable maximum age.
+The Kroxylicious Connection Expiration filter closes client connections after a configurable expiration age.
 This is useful in dynamic environments like Kubernetes where proxy instances scale up and down, helping to rebalance connections across proxy instances.
-For information on using the filter, see the {ConnectionMaxAgeGuide}.
+For information on using the filter, see the {ConnectionExpirationGuide}.
