+++ kroxylicious-docs/docs/_assemblies/assembly-configuring-connection-expiration-filter.adoc
@@ -1,0 +1,39 @@
+:_mod-docs-content-type: ASSEMBLY
+
+// file included in the following:
+//
+// index.adoc
+
+[id='assembly-configuring-connection-expiration-filter-{context}']
+= Configuring the Connection Expiration filter
+
+[role="_abstract"]
+This procedure describes how to set up the Connection Expiration filter by configuring it in Kroxylicious.
+
+.Prerequisites
+
+* An instance of Kroxylicious.
+ifdef::OpenShiftOnly[]
+For information on deploying Kroxylicious, see the {OperatorGuide}.
+endif::OpenShiftOnly[]
+ifndef::OpenShiftOnly[]
+For information on deploying Kroxylicious, see the {ProxyGuide} or {OperatorGuide}.
+endif::OpenShiftOnly[]
+
+.Procedure
+
+. Configure a `ConnectionExpiration` type filter.
+ifdef::include-platform-bare-metal[]
+* In a standalone proxy deployment. See <<con-example-proxy-config-{context}>>
+endif::[]
+ifdef::include-platform-kubernetes[]
+* In a Kubernetes deployment using a `KafkaProtocolFilter` resource. See <<con-example-kafkaprotocolfilter-resource-{context}>>
+endif::[]
+
+ifdef::include-platform-bare-metal[]
+include::../_modules/connection-expiration/con-example-proxy-config.adoc[leveloffset=+1]
+endif::[]
+
+ifdef::include-platform-kubernetes[]
+include::../_modules/connection-expiration/con-example-kafkaprotocolfilter-resource.adoc[leveloffset=+1]
+endif::[]
