--- kroxylicious-docs/docs/_modules/connection-max-age/con-example-kafkaprotocolfilter-resource.adoc
@@ -1,31 +1,0 @@
-:_mod-docs-content-type: CONCEPT
-
-// file included in the following:
-//
-// assembly-configuring-connection-max-age-filter.adoc
-
-[id='con-example-kafkaprotocolfilter-resource-{context}']
-= Example `KafkaProtocolFilter` resource
-
-[role="_abstract"]
-If your instance of Kroxylicious runs on Kubernetes, you must use a `KafkaProtocolFilter` resource to contain the filter configuration.
-
-Here's a complete example of a `KafkaProtocolFilter` resource configured for Connection Max Age:
-
-[source,yaml]
-----
-kind: KafkaProtocolFilter
-metadata:
-  name: my-connection-max-age-filter
-spec:
-  type: ConnectionMaxAge
-  configTemplate:
-    maxAge: "1h"
-    jitter: "5m"
-----
-where:
-
-* `maxAge` (required) is the maximum age of a connection before it will be closed. Must be a positive duration (for example, `"1h"`, `"30m"`, `"3600s"`).
-* `jitter` (optional) is the jitter to apply to the max age, randomizing the effective deadline within `[maxAge - jitter, maxAge + jitter]` per connection. Must be non-negative and not greater than `maxAge`. If omitted, no jitter is applied.
-
-Refer to the {OperatorGuide} for more information about configuration on Kubernetes.
