--- kroxylicious-docs/docs/connection-max-age-guide/index.adoc
@@ -1,24 +1,0 @@
-:experimental:
-include::_assets/attributes.adoc[]
-
-:context: connection-max-age
-:guide: connection-max-age
-
-[id="using-book-{context}"]
-= Connection Max Age Guide
-
-include::_modules/connection-max-age/con-about-{guide}-guide.adoc[leveloffset=+1]
-
-[role="_abstract"]
-This filter closes client connections after a configurable maximum age.
-This is useful in environments like Kubernetes where proxy instances scale up and down.
-Long-lived client connections can become stuck on old proxy instances, causing connection skew.
-By closing connections after a maximum age, clients will reconnect and be balanced across the available proxy instances.
-
-When jitter is configured, the effective deadline is randomized within `[maxAge - jitter, maxAge + jitter]` per connection to avoid thundering herd reconnection storms.
-
-//configuring the connection max age filter
-include::_assemblies/assembly-configuring-connection-max-age-filter.adoc[leveloffset=+1]
-
-//trademark notices
-include::_assets/trademarks.adoc[leveloffset=+1]
