+++ kroxylicious-filters/kroxylicious-connection-expiration/src/main/resources/META-INF/services/io.kroxylicious.proxy.filter.FilterFactory
@@ -1,0 +1,1 @@
+io.kroxylicious.filter.connectionexpiration.ConnectionExpiration
