--- kroxylicious-filters/kroxylicious-connection-max-age/src/main/resources/META-INF/services/io.kroxylicious.proxy.filter.FilterFactory
@@ -1,1 +1,0 @@
-io.kroxylicious.filter.connectionmaxage.ConnectionMaxAgeFilterFactory
