+++ kroxylicious-integration-tests/pom.xml
@@ -115,7 +115,7 @@
         </dependency>
         <dependency>
             <groupId>io.kroxylicious</groupId>
-            <artifactId>kroxylicious-connection-max-age</artifactId>
+            <artifactId>kroxylicious-connection-expiration</artifactId>
             <scope>test</scope>
         </dependency>
         <dependency>
